Create app.js
This JavaScript code defines functions to fetch and display weather data from the OpenWeatherMap API. It allows the user to input a city name, and upon clicking the search button, it fetches the current weather as well as the hourly forecast for the next 24 hours. The data is displayed on the page with temperature, weather description, and an icon representing the weather conditions.